firmware-qcom-boot: add support for CDT file
Qualcomm released the CDT file for the RB3 Gen2. There is a different
CDT per flavor of the device (core, vision, industrial). For now we
only support -core in this layer. The patch is made in such a way that
we can add other flavor when we create the corresponding machines.

The CDT files for RB3 Gen2 are available here:

* https://artifacts.codelinaro.org/artifactory/codelinaro-le/Qualcomm_Linux/QCS6490/cdt/rb3gen2-core-kit.zip
* https://artifacts.codelinaro.org/artifactory/codelinaro-le/Qualcomm_Linux/QCS6490/cdt/rb3gen2-vision-kit.zip
* https://artifacts.codelinaro.org/artifactory/codelinaro-le/Qualcomm_Linux/QCS6490/cdt/rb3gen2-industrial-kit.zip

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>