linux-qcom-staging: add v6.12 kernel recipe
Add a recipe for the linux-qcom-staging 6.12 kernel.

This staging kernel tree includes additional patches on top of the 6.12
kernel, many of which are in the process of being upstreamed.

Starting with 6.12, this tree will continue to catch up with the latest
releases (6.13, 6.14, etc.), with patches being dropped and rebased in
the process. Eventually, all patches from this staging kernel will be
available in the mainline kernel.

Signed-off-by: Viswanath Kraleti <quic_vkraleti@quicinc.com>