linux-firmware: make qcom-qcm6490-wifi package provide -qcs6490-
The package linux-firmware-qcom-qcm6490-wifi also provides a symlink
qcom/qcs6490/wpss.mbn. To follow the pattern established by
linux-firmware-qcom-qcm6490-audio and -compute packages the package
should RPROVIDE the qcs6490 name. Without RPROVIDES it is possible for
the image recipe to attempt installing both package names, resulting in
a conflict between the files provided by the linux-firmware and by the
firmware-qcom-rb3gen2 recipes.

The patch has been submitted to the OE-Core, but until it is merged
provide a local fixup.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>