linux-yocto-dev: qcm6490: dts: pick sc7280 global interrupts patch
This IRQ is used to receive the PCIe controller and link specific events
such as Link Up/Down, MSI, PTM etc... in the driver. So enabling this
IRQ would allow the driver to enumerate the endpoint device and also
retrain the link when the device is removed without user intervention.

Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez@oss.qualcomm.com>