ci: add kas-mirror job to pre-fetch KAS_REPO_REF_DIR just with git
We initially used KAS_REPO_REF_DIR to keep a local copy of the repos mirrors,
to avoid fetching projects from scratch in CI. However, it is implemented such
that the mirrors, once created by kas the first time, are never updated by kas,
they are 'read only'.

Sometimes the fetch from upstream fails and the KAS ended up falling back
to the old content of the mirrors, silently. To fix this we will pre-fetch
all ref front the origin the repos in KAS_REPO_REF_DIR.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>