u-boot: add bbappend to support Android-style boot.img for RB1
This builds a u-boot that can be flashed as boot.img on RB1 and will
load an EFI bootloader from ESP.
An android boot image is generated and deployed based on u-boot.bin and
the matching dtb.

COMPATIBLE_MACHINE is set to 'qcom' in anticipation of future
machines using this chainload method.

Signed-off-by: Koen Kooi <koen.kooi@oss.qualcomm.com>