workflows: Use unique build ids
Build ids on the fileserver simply used the GITHUB_RUN_ID, but it is not
guaranteed to be unique, and other repositories and workflows share the
same tree. Change build ids to be prefixed with the github.repository
name, and suffixed with the github.run_attempt (in case the run has been
given back).

This is still safe from the perspective of directory names and URLs as
GitHub prevents org and repo names to start with a dash, and use only
alphanumeric chars plus single dashes.

Signed-off-by: Loïc Minier <loic.minier@oss.qualcomm.com>