u-boot-qcom: drop recipe
The SRCREV pointed out by the u-boot-qcom recipe does no longer exist in
the repo. Drop the recipe in favour of using u-boot recipe from OE-Core.

Note: this is not a drop-in replacement, since the recipe doesn't
use qtestsign / doesn't create a signed image. It is left as TBD for
anybody using U-Boot in the boot chain.

Cc: Tommaso Merciai <tomm.merciai@gmail.com>
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>