qdl: replace 0.0 in recipe PV with latest release
The most recent qdl release is 2.1, which is 6 months old and we are
referencing a newer revision from git instead.

Still, that git revision is based off a release, so it makes more sense
for PV to start with the revision, so when there is an actual newer
release, possible consumers of the package feed see that there is an
actual update. There is no loss of information, because the source
revision appended after the '+' remains as is.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>