ci: yocto-check-layer: stop using the wrapper
The main task performed by the wrapper is to create a clean dir
to execute the real yocto-check-layer, so we can do the same.

The /bin/bash is because the source command is not provided by /bin/sh.

We also clone the meta-qcom again in the new builddir from the REPO_DIR
to avoid any polution/untracked files that may exist in the repository.
This creates a clean copy of the meta-qcom repo inside the new build dir.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>