image-types-qcom: correct dependency on boot firmware
IMAGE.do_image_qcomflash depends only on .do_rootfs, which in turn
depends only on .do_package_foo tasks. It is possible, that
firmware-qcom-boot-foo.do_deploy (which actually installs boot firmware
into DEPLOY_DIR) is executed in parallel or afterwards, resulting in an
incomplete qcomflash dir.

Add direct dependency on the boot firmware packages in order to make
sure that the qcomflash dir is complete.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>