linux-qcom-next: avoid fetcher errors by using 'nobranch' in SRC_URI
The `qcom-next` kernel branch is retargeted on every tag upmerge, which
results in the loss of commit history. However, the tagged commits are
preserved in the repo and can still be accessed via their commit IDs.

When the `branch` parameter is specified in `SRC_URI`, BitBake's fetcher
restricts the `SRCREV` lookup to that branch. This causes fetch failures
if the recipe still points to an older tag while `qcom-next`has moved on
to a new one.

To avoid such issues, use `nobranch=1` in `SRC_URI`. This allows BitBake
to search for the specified `SRCREV` across the repository, including
commits that are no longer associated with the current branch tip but are
still preserved by tag.

The `branch` parameter should be used only when working with `devupstream`,
where a stable branch reference is expected.

Signed-off-by: Viswanath Kraleti <viswanath.kraleti@oss.qualcomm.com>