ci: reduce the amount of files published for each build
Each build exports close to 590 files, because it was initially
decided to 'blindly' publish the deploy_dir folder which is bitbake
output directory.

In a typical build, there are 25 symlinks which end up being duplicate
files on the file server (1.3GB per platform). The symlinks are
convenient/shortened filenames, and don't need to be published.

deploy_dir also contains 'partitions' folder which is an intermediate
artifact used to generate the image GPT files, with 490 files.

To clean up the published files, and to avoid growing it by mistake in
the future, this commit switches the logic. Instead of publishing the
entire deploy_dir folder (with minimal cleanup), the CI job now only
publishes real files (no symlinks) located in the root of
deploy_dir. Any additional files needed (later) can be manually added.

For a local RB3G2 build, it's 26 files, for 900MB.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>