ci: correctly copy symlinks for boot  and ext4 images
The find tool takes an expression as an argument, with different parts
of expression (including actions such as -exec) being tied by logical
operators, -o and -a, with -a being a default. Thus the command line
that was added in 811432bc1867 ("ci: include symlink for files used by
LAVA tests") is interpreted as 'find (at depth 1 and symlinks and boot
images) or (rootfs files) or (qcomflash archives and copy them), which
resulted in CI publishing qcomflash.tar.gz archives, but not adding
symlinks for the boot images nor for the rootfs compressed images.

Add parentheses to the find command line so that it is interpreted as
'find at depth 1 and symlinks and (boot images or rootfs images or
qcomflash archives) and copy them'.

Fixes: 811432bc1867 ("ci: include symlink for files used by LAVA tests")
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>