linux-qcom-next: move to 6.16-20250801 tag
Migrating qcom-next to 6.16 is a major release bump. Ideally, the recipe
should be versioned as 6.16. However, since the current recipe version
includes rcX, using 6.16 directly would trigger a [version-going-backwards]
error. To avoid this, 6.16-rel is opted as the interim version.

When migrating to 6.17-rc1, the version will be updated to 6.16-rel+6.17-rc1.
Finally, at the time of the 6.17 release, the version will be set to 6.17,
and versioning will align normally from that point onward.

Signed-off-by: Viswanath Kraleti <viswanath.kraleti@oss.qualcomm.com>