image-types-qcom: respect the IMAGE_NAME_SUFFIX variable
IMAGE_NAME_SUFFIX [1]: Suffix used for the image output filename defaults to ".rootfs" to distinguish
the image file from other files created during image building; however if this suffix is redundant or not
desired you can clear the value of this variable (set the value to “”). For example, this is typically
cleared in Initramfs image recipes.

[1] https://docs.yoctoproject.org/ref-manual/variables.html#term-IMAGE_NAME_SUFFIX

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>