ci: clean up OE persistent storage from time to time
sstate folder can grow significantly over time, and needs to be
cleaned up. The simplest thing to do is to remove it regularly. To
avoid any manual intervention, we append '-year-month' to the folder
name, and we delete older folders. We keep current and last month to
avoid deleting the 'last' month folder while it might still be in use
by current running builds (around the first day of each month).

Closes #1018.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>