ci: run a full sanity build first
The matrix jobs results in a large amount of jobs being started at the
same time. Running all jobs in parallel is not ideal to leverage the
OE sstate cache. For any common tasks (native, armv8, .. ) when the
object is missing from the sstate cache, it's likely going to be a
miss for several jobs. As a result, each of them will build the missed
object, and each of them will attempt to create the object in the
sstate cache.

Instead, with this change, we first build for the generic qcom_armv8
machine (which uses armv8a tune) and rb3 gen2 (which uses armv8-2a
tune), and if that builds succeeds we carry on with the other
machines, with the hope that the sstate cache hit will be better.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>