ci: add world.yml KAS configuration snippet
We want to be able to run a 'meta-qcom world' build, which means try
to build all recipes available in meta-qcom. This can be useful for
CI/testing purpose. Adding world.yml snippet will run such a
meta-qcom world build.

We will extend to include meta-qcom-distro eventually, but for now
adding it breaks the world build.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>