CI: drop useless page_owner=on kernel arg
The kernel complains about the unknown parameter:

Unknown kernel command line parameters "page_owner=on", will be passed to user space.

This warning makes sense, as page_owner=on is only available if
CONFIG_PAGE_OWNER is also enabled. As such, it doesn't make sense to
have this param in the default configuration: to make it useful it would
be required to rebuild the kernel anyway.

Drop it from the CI config.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>