packagegroup-qcom-ci: remove recipe
This packagegroup was created to ensure that we are building extra
recipes in CI which are not included in any image. Now that we have
the support for meta-qcom local world build, this packagegroup is no
longer needed, since the world build would take care of it (and even
more).

Also remove packagegroup-qcom-graphics whose sole existence was to be
included in the CI packagegroup to ensure it gets built in CI, which
is no longer needed with the world build.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>