qcom-armv8: add efi machine_feature
Up until now, we had assumed to enable EFI/ESP  for all machines
but qcom-armv7. While it worked, it's more appropriate to rely on
'efi' MACHINE_FEATURE directly.

We are about to enable ESP only for machines with efi support, which
will disable ESP for qcom-armv8, so we need to enable efi in this
machine.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>