qcom-adreno: update KGSL tip and drop merged patches from recipe
Update the SRCREV to point to the latest commit in the KGSL source code
repository. This update brings in several improvements and fixes:

- Fix multiple warnings
- Fix compilation errors in the driver with standard 6.12 kernel
- Use secure firmware with mbn format
- Add support to check /lib/firmware/qcom for SQE and GMU firmwares
- Conditionally invoke set_supported_hw API based on speedbin fuse
- Add support for standard compatible strings
- Ensure IRQ name string persists using devm_kasprintf
- Avoid use after free in kgsl_destroy_ion()

Additionally, remove patches that have become redundant, as
their changes have already been merged in the source code and no longer
required to be applied explicitly in kgsl-dlkm recipe.

Signed-off-by: Kaushal Sanadhya <quic_ksanadhy@quicinc.com>